updated scoring system +1.5 correct -1 incorrect and showing on both pages and works on 4s and up

+++ YCJAPViewController.h
@@ -18,6 +18,6 @@
 @property (weak, nonatomic) IBOutlet UIButton *nextButton;
 @property (weak, nonatomic) IBOutlet UIButton *contribute;
 @property (weak, nonatomic) IBOutlet UILabel *scoreLabel;
-@property  int score;
+@property  double score;
 
 @end
